hardware: bill decimal places modify scaling factor
Lesson learned: MDB doc saying "decimal places indicates the number on display" actually does more than just display.

+++ hardware/mdb/config/mdb_config.go
@@ -4,7 +4,7 @@ package mdb_config
 
 type Config struct { //nolint:maligned
 	Bill struct {
-		ScalingFactor uint16 `hcl:"scaling_factor"`
+		ScalingFactor int `hcl:"scaling_factor"`
 	}
 	Coin struct { //nolint:maligned
 		DispenseSmart      bool `hcl:"dispense_smart"`
